+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Fixed bug in `analyze_results!` functions where if some parts of the algorithm was not run, analyzing results would error
 - Fixed stats unit test where expected variance was too close to the allowed tolerance
 - Fixed bug in `get_timestep_fault_currents` where there was a typo in the variable name for phase currents
 - Added commandline argument `opt-disp-solver` to select which solver to use for optimal dispatch